Implementation/requirements.txt
matplotlib==3.6.3
mlxtend==0.21.0
numpy==1.23.5
pandas==1.5.2
Pillow==9.4.0
Pillow==10.4.0
scikit_learn==1.2.1
seaborn==0.13.2
tensorflow==2.9.0
